requirements.txt
fastapi==0.114.1
uvicorn==0.30.5
streamlit==1.39.0
PyPDF2==3.0.1
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
openai>=2.8.0
python-dotenv==1.2.1
requests==2.32.3
python-multipart==0.0.20
tf-keras>=2.20.0
